+++ lib/src/bpqueue.cpp
@@ -8,6 +8,6 @@
 // template <typename T, typename Int, class Container>
 // dllink<std::pair<T, Int>> bpqueue<T, Int, Container>::sentinel {};
 
-template class bpqueue<int, int16_t>;
-// template class bpqueue<int, int16_t,
-//                FMPmr::vector<dllink<std::pair<int, int16_t>> > >;
+template class bpqueue<int, int32_t>;
+// template class bpqueue<int, int32_t,
+//                FMPmr::vector<dllink<std::pair<int, int32_t>> > >;
